+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/pokemon/gen1/CuboneModel.kt
@@ -23,8 +23,8 @@ class CuboneModel(root: ModelPart) : PokemonPoseableModel(), HeadedFrame, BipedF
     override val rootPart = root.registerChildWithAllChildren("cubone")
     override val head = getPart("head")
 
-    override val leftLeg = getPart("leftleg")
-    override val rightLeg = getPart("rightleg")
+    override val leftLeg = getPart("left_leg")
+    override val rightLeg = getPart("right_leg")
 
     override val portraitScale = 2.0F
     override val portraitTranslation = Vec3d(0.1, -0.7, 0.0)
@@ -36,13 +36,16 @@ class CuboneModel(root: ModelPart) : PokemonPoseableModel(), HeadedFrame, BipedF
     lateinit var walk: PokemonPose
 
     override fun registerPoses() {
+        val blink = quirk("blink") { bedrockStateful("cubone", "blink").setPreventsIdle(false) }
+
         standing = registerPose(
             poseName = "standing",
             poseTypes = STATIONARY_POSES + UI_POSES,
             transformTicks = 10,
+            quirks = arrayOf(blink),
             idleAnimations = arrayOf(
                 singleBoneLook(),
-                bedrock("cubone", "ground_idle")
+                bedrock("cubone", "idle")
             )
         )
 
@@ -50,10 +53,11 @@ class CuboneModel(root: ModelPart) : PokemonPoseableModel(), HeadedFrame, BipedF
             poseName = "walk",
             poseTypes = MOVING_POSES,
             transformTicks = 10,
+            quirks = arrayOf(blink),
             idleAnimations = arrayOf(
                 BipedWalkAnimation(this, periodMultiplier = 1.1F, amplitudeMultiplier = 1.15f),
                 singleBoneLook(),
-                bedrock("cubone", "ground_idle")
+                bedrock("cubone", "idle")
                 //bedrock("cubone", "ground_walk")
             )
         )