+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/pokemon/gen1/MarowakModel.kt
@@ -23,8 +23,8 @@ class MarowakModel(root: ModelPart) : PokemonPoseableModel(), HeadedFrame, Biped
     override val rootPart = root.registerChildWithAllChildren("marowak")
     override val head = getPart("head")
 
-    override val leftLeg = getPart("leftleg")
-    override val rightLeg = getPart("rightleg")
+    override val leftLeg = getPart("left_leg")
+    override val rightLeg = getPart("right_leg")
 
     override val portraitScale = 1.7F
     override val portraitTranslation = Vec3d(0.0, 0.15, 0.0)
@@ -36,22 +36,26 @@ class MarowakModel(root: ModelPart) : PokemonPoseableModel(), HeadedFrame, Biped
     lateinit var walk: PokemonPose
 
     override fun registerPoses() {
+        val blink = quirk("blink") { bedrockStateful("venomoth", "blink").setPreventsIdle(false) }
+
         standing = registerPose(
             poseName = "standing",
             poseTypes = STATIONARY_POSES + UI_POSES,
+            quirks = arrayOf(blink),
             idleAnimations = arrayOf(
                 singleBoneLook(),
-                bedrock("marowak", "ground_idle")
+                bedrock("marowak", "idle")
             )
         )
 
         walk = registerPose(
             poseName = "walk",
             poseTypes = MOVING_POSES,
+            quirks = arrayOf(blink),
             idleAnimations = arrayOf(
                 BipedWalkAnimation(this, periodMultiplier = 0.9F, amplitudeMultiplier = 1.1f),
                 singleBoneLook(),
-                bedrock("marowak", "ground_idle")
+                bedrock("marowak", "idle")
                 //bedrock("marowak", "ground_walk")
             )
         )