+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/pokemon/gen2/YanmaModel.kt
@@ -27,7 +27,7 @@ class YanmaModel(root: ModelPart) : PokemonPoseableModel(), HeadedFrame {
     override val head = getPart("head")
 
     override val portraitScale = 2.1F
-    override val portraitTranslation = Vec3d(-0.6, -0.4, 0.0)
+    override val portraitTranslation = Vec3d(-0.6, -1.0, 0.0)
 
     override val profileScale = 0.75F
     override val profileTranslation = Vec3d(0.0, 0.5, 0.0)